# businessPartner/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Business Partner Service",
    "description": "API Documentation for Business Partner Service",
    "version": "0.1-oas3"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/bp/mentors": {
      "get": {
        "responses": {
          "200": {
            "description": "Returns list of available Franchise Mentor Assignment"
          }
        }
      },
      "put": {
        "summary": "Assign a mentor to a business partner",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Message"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "payload": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              },
              "encoding": {
                "payload": {
                  "contentType": "application/json"
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Mentor Assigned to Franchise"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Message": {
        "title": "Assignment object",
        "type": "object",
        "properties": {
          "BPID": {
            "type": "string",
            "description": "Business Partner ID/ Franchise ID"
          },
          "MENTORID": {
            "type": "string",
            "description": "Mentor ID"
          }
        },
        "required": [
          "text"
        ]
      }
    }
  }
}